feat: move grammar types/conversions to schema, add 0.6→0.7 migration
