+++ packages/base/src/__tests__/migrations.spec.ts
@@ -10,6 +10,11 @@
 import fs from 'fs';
 import path from 'path';
 
+// Stub UUID so grammar migration fixtures contain deterministic rule ids.
+jest.mock('@lumino/coreutils', () => ({
+  UUID: { uuid4: () => 'test-uuid' },
+}));
+
 import { migrateDocument } from '../../../schema/src/migrations';
 
 const FIXTURES_ROOT = path.resolve(
@@ -17,7 +22,10 @@ const FIXTURES_ROOT = path.resolve(
   '../../../../packages/schema/test-fixtures/migrations',
 );
 
-const VERSION_PAIRS: [string, string][] = [['v0.5.0', 'v0.6.0']];
+const VERSION_PAIRS: [string, string][] = [
+  ['v0.5.0', 'v0.6.0'],
+  ['v0.6.0', 'v0.7.0'],
+];
 
 function sharedFixtureNames(fromV: string, toV: string): string[] {
   const fromDir = path.join(FIXTURES_ROOT, fromV);
