+++ packages/schema/src/schema/project/jgis.json
@@ -6,7 +6,7 @@
   "properties": {
     "schemaVersion": {
       "type": "string",
-      "default": "0.6.0"
+      "default": "0.7.0"
     },
     "layers": {
       "$ref": "#/definitions/jGISLayers"
